# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blood Bank API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Returns User Credentials",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "loginData": {
                                            "type": "object",
                                            "example": "id: 1,name:saransh, email:admin@admin.com, age:22, address:e-15, city:delhi,state: delhi, gender:male, phone:123, blood_type:A+, token: token"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "description": "Returns User Data",
                "requestBody": {
                    "description": "Provide provide all details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "admin.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "address"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Delhi"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "delhi"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 987654321
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "A+"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "registerData": {
                                            "type": "object",
                                            "example": "id: 1,name:admin, email:admin@admin.com, age:22, address:e-15, city:delhi,state: delhi, gender:male, phone:123, blood_type:A+"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout",
                "description": "Logout current user",
                "requestBody": {
                    "description": "No parameter required",
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "logoutData": {
                                            "type": "object",
                                            "example": "Successfully Logged Out"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update User Details",
                "description": "Returns User Data",
                "requestBody": {
                    "description": "Provide provide only parameter you want to update,",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "admin.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "address"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Delhi"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "delhi"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 987654321
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "A+"
                                    },
                                    "oldPassword": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updatedData": {
                                            "type": "object",
                                            "example": "id: 1,name:admin, email:admin@admin.com, age:22, address:e-15, city:delhi,state: delhi, gender:male, phone:123, blood_type:A+"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/findDonors": {
            "post": {
                "tags": [
                    "Donor"
                ],
                "summary": "Get Donors",
                "description": "Get all donors, get donors by State, City, Blood Group",
                "requestBody": {
                    "description": "Provide only parameters to search, except enter 'all'",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "state": {
                                        "type": "string",
                                        "example": "All"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "All"
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "All"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donorsData": {
                                            "type": "object",
                                            "example": "[{id: 1,name:admin, email:admin@admin.com, age:22, address:e-15, city:delhi,state: delhi, gender:male, phone:123, blood_type:A+}]"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/getAllBloodType": {
            "post": {
                "tags": [
                    "Donor"
                ],
                "summary": "Get All Blood Type",
                "description": "Get all blood Type",
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "bloodTypeData": {
                                            "type": "object",
                                            "example": "[A+,A-,A1+,A1-,A1B+,A1B-,A2+,A2-,A2B+,A2B-,AB+,AB-,B+,B-,Bombay Blood Group,INRA,O+,O-]"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/getAllCountries": {
            "post": {
                "tags": [
                    "Geographical"
                ],
                "summary": "Get all countries",
                "description": "Returns All Countries Data",
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": "[{id: 1,name:India, phonecode:+91}]"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "sanctum": [],
        "0": []
    }
}